fix(master) send proper email alerts for TrueNAS HA customers
